requirements.txt
# ============================================================
# SimCLR Project Dependencies
# ============================================================
# Core: PyTorch with CUDA 12.4 support
# Install: pip install -r requirements.txt
# ============================================================

# Deep Learning
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Data & Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Dimensionality Reduction (optional: UMAP)
umap-learn>=0.5.0

# MLOps
mlflow>=2.10.0
dvc>=3.40.0

# Image Processing
Pillow>=10.0.0

# Interactive Demo
streamlit>=1.30.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0

# Code Quality (dev)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Testing (dev)
pytest>=7.4.0
